refactor(ui): extract weapons / modifiers / notes / equipment tab mixins
The final four per-tab mixins:

* WeaponsTabMixin -> Tab 7 (build_ui_page_7, melee + ranged + arrows).
* ModifiersTabMixin -> Tab 8 (build_ui_page_8).
* NotesTabMixin -> Tab 9 (build_ui_page_9, including build_pers_info
  with self.pers_info_widgets and the .link aliases).
* EquipmentTabMixin -> Tab 10 (build_ui_page_10, equipment list +
  MoneyWidget).

main.py: 753 -> 533 lines.
