+++ l5r/ui/tabs/weapons.py
@@ -1,0 +1,75 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2014-2026 Daniele Simonetti
+#
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation; either version 2 of the License, or
+# (at your option) any later version.
+#
+# Tab 7 — Weapons (melee + ranged + arrows). Extracted from l5r/main.py
+# during the Phase 4 split — no behaviour changes.
+
+from qtpy import QtGui
+
+import l5r.models as models
+import l5r.widgets as widgets
+
+from l5r.util.fsutil import get_icon_path
+
+
+class WeaponsTabMixin:
+    """Tab 7: three weapon tables with shared add/edit/remove toolbar."""
+
+    def build_ui_page_7(self):
+        self.melee_view_model = models.WeaponTableViewModel('melee', self)
+        self.ranged_view_model = models.WeaponTableViewModel('ranged', self)
+        self.arrow_view_model = models.WeaponTableViewModel('arrow', self)
+
+        def _make_sortable(model):
+            # enable sorting through a proxy model
+            sort_model_ = models.ColorFriendlySortProxyModel(self)
+            sort_model_.setDynamicSortFilter(True)
+            sort_model_.setSourceModel(model)
+            return sort_model_
+
+        # weapon vertical toolbar
+        def _make_vertical_tb(has_custom, has_edit, has_qty, filt):
+            vtb = widgets.VerticalToolBar(self)
+            vtb.setProperty('filter', filt)
+            vtb.addStretch()
+            vtb.addButton(QtGui.QIcon(get_icon_path('buy', (16, 16))),
+                          self.tr("Add weapon"), self.sink3.show_add_weapon)
+            if has_custom:
+                vtb.addButton(QtGui.QIcon(get_icon_path('custom', (16, 16))),
+                              self.tr("Add custom weapon"), self.sink3.show_add_cust_weapon)
+            if has_edit:
+                vtb.addButton(QtGui.QIcon(get_icon_path('edit', (16, 16))),
+                              self.tr("Edit weapon"), self.sink3.edit_selected_weapon)
+            vtb.addButton(QtGui.QIcon(get_icon_path('minus', (16, 16))),
+                          self.tr("Remove weapon"), self.sink3.remove_selected_weapon)
+            if has_qty:
+                vtb.addButton(QtGui.QIcon(get_icon_path('add', (16, 16))),
+                              self.tr("Increase Quantity"), self.sink3.on_increase_item_qty)
+                vtb.addButton(QtGui.QIcon(get_icon_path('minus', (16, 16))),
+                              self.tr("Decrease Quantity"), self.sink3.on_decrease_item_qty)
+
+            vtb.addStretch()
+            return vtb
+
+        melee_vtb = _make_vertical_tb(True, True, False, 'melee')
+        ranged_vtb = _make_vertical_tb(True, True, False, 'ranged')
+        arrow_vtb = _make_vertical_tb(False, False, True, 'arrow')
+
+        models_ = [
+            (self.tr("Melee Weapons"), 'table', _make_sortable(self.melee_view_model), None, melee_vtb, None),
+            (self.tr("Ranged Weapons"), 'table', _make_sortable(self.ranged_view_model), None, ranged_vtb, None),
+            (self.tr("Arrows"), 'table', _make_sortable(self.arrow_view_model), None, arrow_vtb, None)
+        ]
+
+        frame_, views_ = self._build_generic_page(models_)
+
+        melee_vtb .setProperty('source', views_[0])
+        ranged_vtb.setProperty('source', views_[1])
+        arrow_vtb .setProperty('source', views_[2])
+
+        self.tabs.addTab(frame_, self.tr("Weapons"))
